+++ org.knime.knip.patents/src/org/knime/knip/patents/util/nodes/drawings/OPSImagesNodeFactory.xml
@@ -5,8 +5,8 @@
 	Contributors: IBM Corporation - initial API and implementation -->
 
 <!DOCTYPE knimeNode PUBLIC "-//UNIKN//DTD KNIME Node 2.0//EN" "http://www.knime.org/Node.dtd">
-<knimeNode icon="icons/logo.png" type="Manipulator">
-	<name>OPS Images</name>
+<knimeNode icon="../logo.png" type="Manipulator">
+	<name>OPS Drawings</name>
 	<shortDescription>Download drawings of a patent.</shortDescription>
 
 	<fullDescription>
