+++ org.knime.knip.patents/src/org/knime/knip/patents/util/nodes/drawings/OPSImagesNodeModel.java
@@ -23,13 +23,13 @@ import org.knime.core.data.def.StringCell;
 import org.knime.core.node.NodeLogger;
 import org.knime.core.util.Pair;
 import org.knime.knip.patents.KNIMEOPSPlugin;
-import org.knime.knip.patents.util.AbstractOPSModel;
 import org.knime.knip.patents.util.AccessTokenGenerator;
+import org.knime.knip.patents.util.nodes.AbstractOPSNodeModel;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.NodeList;
 
-public class OPSImagesNodeModel extends AbstractOPSModel {
+public class OPSImagesNodeModel extends AbstractOPSNodeModel {
 
 	private static final NodeLogger LOGGER = NodeLogger
 			.getLogger(OPSImagesNodeModel.class);
