+++ org.knime.knip.patents/src/org/knime/knip/patents/util/nodes/fullcycle/FullCycleNodeFactory.xml
@@ -5,8 +5,8 @@
 	Contributors: IBM Corporation - initial API and implementation -->
 
 <!DOCTYPE knimeNode PUBLIC "-//UNIKN//DTD KNIME Node 2.0//EN" "http://www.knime.org/Node.dtd">
-<knimeNode icon="icons/logo.png" type="Manipulator">
-	<name>FullCycle</name>
+<knimeNode icon="../logo.png" type="Manipulator">
+	<name>OPS FullCycle</name>
 	<shortDescription>Downloads the full cycle information of a patent.</shortDescription>
 
 	<fullDescription>
